src/docs/palmvein_backend/requirements.txt

flask==2.2.3
flask-cors==3.0.10
opencv-python==4.7.0.72
mediapipe==0.9.1.0
numpy==1.24.2
pyjwt==2.6.0
bcrypt==4.0.1
python-dotenv==1.0.0
gunicorn==20.1.0
